# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CaseOut": {
        "additionalProperties": false,
        "properties": {
          "case_number": {
            "title": "Case Number",
            "type": "string"
          },
          "close_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Reason"
          },
          "closed_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed By"
          },
          "conflict_flag": {
            "title": "Conflict Flag",
            "type": "boolean"
          },
          "conflict_match_result_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflict Match Result Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "duration_months": {
            "title": "Duration Months",
            "type": "integer"
          },
          "end_date": {
            "format": "date",
            "title": "End Date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "match_result_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Result Id"
          },
          "provider_id": {
            "title": "Provider Id",
            "type": "string"
          },
          "sanction_record_id": {
            "format": "uuid",
            "title": "Sanction Record Id",
            "type": "string"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "case_number",
          "provider_id",
          "sanction_record_id",
          "match_result_id",
          "status",
          "start_date",
          "end_date",
          "duration_months",
          "created_by",
          "closed_by",
          "close_reason",
          "conflict_flag",
          "conflict_match_result_id",
          "created_at"
        ],
        "title": "CaseOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_ProviderListItemOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProviderListItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ProviderListItemOut]",
        "type": "object"
      },
...
      "ProviderDetailOut": {
        "additionalProperties": false,
        "properties": {
          "address_line1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line1"
          },
          "address_line2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line2"
          },
          "cases": {
            "items": {
              "$ref": "#/components/schemas/CaseOut"
            },
            "title": "Cases",
            "type": "array"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "compliance_status": {
            "description": "`EXCLUDED`: an active case. `UNDER_REVIEW`: the engine's choice on a current result a reviewer has not decided. `CLEAR`: neither.",
            "enum": [
              "EXCLUDED",
              "UNDER_REVIEW",
              "CLEAR"
            ],
            "title": "Compliance Status",
            "type": "string"
          },
          "dba_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dba Name"
          },
          "dob": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dob"
          },
          "ein": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ein"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "is_organization": {
            "title": "Is Organization",
            "type": "boolean"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "license_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Number"
          },
          "license_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License State"
          },
          "matches": {
            "description": "Results that ranked this provider, newest first, superseded ones included.",
            "items": {
              "$ref": "#/components/schemas/ProviderMatchOut"
            },
            "title": "Matches",
            "type": "array"
          },
          "middle_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Middle Name"
          },
          "npi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Npi"
          },
          "organization_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Name"
          },
          "provider_id": {
            "title": "Provider Id",
            "type": "string"
          },
          "specialty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialty"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "suffix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suffix"
          },
          "zip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zip"
          }
        },
        "required": [
          "provider_id",
          "npi",
          "first_name",
          "middle_name",
          "last_name",
          "suffix",
          "organization_name",
          "dba_name",
          "ein",
          "is_organization",
          "dob",
          "address_line1",
          "address_line2",
          "city",
          "state",
          "zip",
          "license_number",
          "license_state",
          "specialty",
          "status",
          "compliance_status",
          "cases",
          "matches"
        ],
        "title": "ProviderDetailOut",
        "type": "object"
      },
      "ProviderListItemOut": {
        "additionalProperties": false,
        "properties": {
          "address_line1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line1"
          },
          "address_line2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line2"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "compliance_status": {
            "description": "`EXCLUDED`: an active case. `UNDER_REVIEW`: the engine's choice on a current result a reviewer has not decided. `CLEAR`: neither.",
            "enum": [
              "EXCLUDED",
              "UNDER_REVIEW",
              "CLEAR"
            ],
            "title": "Compliance Status",
            "type": "string"
          },
          "dba_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dba Name"
          },
          "dob": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dob"
          },
          "ein": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ein"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "is_organization": {
            "title": "Is Organization",
            "type": "boolean"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "license_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Number"
          },
          "license_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License State"
          },
          "middle_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Middle Name"
          },
          "npi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Npi"
          },
          "organization_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Name"
          },
          "provider_id": {
            "title": "Provider Id",
            "type": "string"
          },
          "specialty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialty"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "suffix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suffix"
          },
          "zip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zip"
          }
        },
        "required": [
          "provider_id",
          "npi",
          "first_name",
          "middle_name",
          "last_name",
          "suffix",
          "organization_name",
          "dba_name",
          "ein",
          "is_organization",
          "dob",
          "address_line1",
          "address_line2",
          "city",
          "state",
          "zip",
          "license_number",
          "license_state",
          "specialty",
          "status",
          "compliance_status"
        ],
        "title": "ProviderListItemOut",
        "type": "object"
      },
      "ProviderMatchOut": {
        "additionalProperties": false,
        "description": "A decision in which this provider was a candidate.",
        "properties": {
          "approved_provider_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved Provider Id"
          },
          "calibrated_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibrated Confidence"
          },
          "candidate_posterior": {
            "title": "Candidate Posterior",
            "type": "number"
          },
          "candidate_rank": {
            "title": "Candidate Rank",
            "type": "integer"
          },
          "chosen_provider_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chosen Provider Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_organization": {
            "title": "Is Organization",
            "type": "boolean"
          },
          "posterior": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posterior"
          },
          "raw_match_weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Match Weight"
          },
          "record_id": {
            "title": "Record Id",
            "type": "string"
          },
          "review_status": {
            "title": "Review Status",
            "type": "string"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "reviewer_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Comment"
          },
          "route": {
            "title": "Route",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "sanction_record_id": {
            "format": "uuid",
            "title": "Sanction Record Id",
            "type": "string"
          },
          "sanction_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sanction Type"
          },
          "source_authority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Authority"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "subject_name": {
            "title": "Subject Name",
            "type": "string"
          },
          "superseded_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Superseded By"
          }
        },
        "required": [
          "id",
          "run_id",
          "sanction_record_id",
          "decision",
          "chosen_provider_id",
          "approved_provider_id",
          "posterior",
          "calibrated_confidence",
          "raw_match_weight",
          "route",
          "review_status",
          "reviewed_by",
          "reviewed_at",
          "reviewer_comment",
          "superseded_by",
          "created_at",
          "record_id",
          "source_authority",
          "subject_name",
          "is_organization",
          "state",
          "sanction_type",
          "candidate_rank",
          "candidate_posterior"
        ],
        "title": "ProviderMatchOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Provider sanctions and exclusions reconciliation.\n\nMatches sanction records against a provider master file with a Fellegi-Sunter\nengine, EM-learned weights and isotonic calibration, and routes the grey band to\nan LLM adjudicator. Every run is pinned to a data snapshot and can be replayed.",
    "title": "Concordance",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/providers": {
      "get": {
        "description": "The directory, each provider beside its derived compliance status.",
        "operationId": "list_providers_providers_get",
        "parameters": [
          {
            "description": "Page size.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size.",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Rows to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "A name substring, an exact provider id, or a 10-digit NPI.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A name substring, an exact provider id, or a 10-digit NPI.",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2,
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "specialty",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Specialty"
            }
          },
          {
            "in": "query",
            "name": "record_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "individual",
                    "organization"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Record Type"
            }
          },
          {
            "in": "query",
            "name": "compliance",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "EXCLUDED",
                    "UNDER_REVIEW",
                    "CLEAR"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compliance"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "provider_id",
              "enum": [
                "provider_id",
                "name",
                "state"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProviderListItemOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Providers",
        "tags": [
          "providers"
        ]
      }
    },
    "/providers/{provider_id}": {
      "get": {
        "description": "One provider: the full record, every case, and every result that ranked them.",
        "operationId": "get_provider_providers__provider_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Provider",
        "tags": [
          "providers"
        ]
      }
    },
...
  }
}